requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
ultralytics>=8.3.0
pillow>=10.0.0
numpy>=1.26.0
huggingface_hub>=0.24.0
